code organization update
