+++ src/agdb/query/query_values.rs
@@ -1,8 +1,6 @@
-use super::query_ids::QueryIds;
 use crate::db::db_key_value::DbKeyValue;
 
 pub enum QueryValues {
-    Ids(QueryIds),
     Single(Vec<DbKeyValue>),
     Multi(Vec<Vec<DbKeyValue>>),
 }
