--- src/agdb/query_builder/insert_values_ids.rs
@@ -1,9 +1,0 @@
-use crate::query::insert_values_query::InsertValuesQuery;
-
-pub struct InsertValuesIds(pub InsertValuesQuery);
-
-impl InsertValuesIds {
-    pub fn query(self) -> InsertValuesQuery {
-        self.0
-    }
-}
