+++ src/agdb/query_builder/remove_values.rs
@@ -1,10 +1,11 @@
-use super::remove_values_ids::RemoveValuesIds;
 use crate::query::query_id::QueryId;
 use crate::query::query_ids::QueryIds;
 use crate::query::remove_values_query::RemoveValuesQuery;
 use crate::query::search_query::SearchQuery;
 
 pub struct RemoveValues(pub RemoveValuesQuery);
+
+pub struct RemoveValuesIds(pub RemoveValuesQuery);
 
 impl RemoveValues {
     pub fn id(mut self, id: QueryId) -> RemoveValuesIds {
@@ -23,5 +24,11 @@ impl RemoveValues {
         self.0 .0.ids = QueryIds::Search(query);
 
         RemoveValuesIds(self.0)
+    }
+}
+
+impl RemoveValuesIds {
+    pub fn query(self) -> RemoveValuesQuery {
+        self.0
     }
 }
