--- src/agdb/query_builder/remove_values_ids.rs
@@ -1,9 +1,0 @@
-use crate::query::remove_values_query::RemoveValuesQuery;
-
-pub struct RemoveValuesIds(pub RemoveValuesQuery);
-
-impl RemoveValuesIds {
-    pub fn query(self) -> RemoveValuesQuery {
-        self.0
-    }
-}
