--- src/agdb/query_builder/select_aliases_ids.rs
@@ -1,9 +1,0 @@
-use crate::query::select_aliases_query::SelectAliasesQuery;
-
-pub struct SelectAliasesIds(pub SelectAliasesQuery);
-
-impl SelectAliasesIds {
-    pub fn query(self) -> SelectAliasesQuery {
-        self.0
-    }
-}
