+++ src/agdb/query_builder/select_key_count.rs
@@ -1,10 +1,11 @@
-use super::select_key_count_ids::SelectKeyCountIds;
 use crate::query::query_id::QueryId;
 use crate::query::query_ids::QueryIds;
 use crate::query::search_query::SearchQuery;
 use crate::query::select_key_count_query::SelectKeyCountQuery;
 
 pub struct SelectKeyCount(pub SelectKeyCountQuery);
+
+pub struct SelectKeyCountIds(pub SelectKeyCountQuery);
 
 impl SelectKeyCount {
     pub fn id(mut self, id: QueryId) -> SelectKeyCountIds {
@@ -23,5 +24,11 @@ impl SelectKeyCount {
         self.0 .0 = QueryIds::Search(query);
 
         SelectKeyCountIds(self.0)
+    }
+}
+
+impl SelectKeyCountIds {
+    pub fn query(self) -> SelectKeyCountQuery {
+        self.0
     }
 }
