--- src/agdb/query_builder/select_key_count_ids.rs
@@ -1,9 +1,0 @@
-use crate::query::select_key_count_query::SelectKeyCountQuery;
-
-pub struct SelectKeyCountIds(pub SelectKeyCountQuery);
-
-impl SelectKeyCountIds {
-    pub fn query(self) -> SelectKeyCountQuery {
-        self.0
-    }
-}
