+++ src/agdb/query_builder/select_keys.rs
@@ -1,10 +1,11 @@
-use super::select_keys_ids::SelectKeysIds;
 use crate::query::query_id::QueryId;
 use crate::query::query_ids::QueryIds;
 use crate::query::search_query::SearchQuery;
 use crate::query::select_keys_query::SelectKeysQuery;
 
 pub struct SelectKeys(pub SelectKeysQuery);
+
+pub struct SelectKeysIds(pub SelectKeysQuery);
 
 impl SelectKeys {
     pub fn id(mut self, id: QueryId) -> SelectKeysIds {
@@ -23,5 +24,11 @@ impl SelectKeys {
         self.0 .0 = QueryIds::Search(query);
 
         SelectKeysIds(self.0)
+    }
+}
+
+impl SelectKeysIds {
+    pub fn query(self) -> SelectKeysQuery {
+        self.0
     }
 }
