--- src/agdb/query_builder/select_keys_ids.rs
@@ -1,9 +1,0 @@
-use crate::query::select_keys_query::SelectKeysQuery;
-
-pub struct SelectKeysIds(pub SelectKeysQuery);
-
-impl SelectKeysIds {
-    pub fn query(self) -> SelectKeysQuery {
-        self.0
-    }
-}
