--- src/agdb/query_builder/select_limit.rs
@@ -1,9 +1,0 @@
-use crate::query::search_query::SearchQuery;
-
-pub struct SelectLimit(pub SearchQuery);
-
-impl SelectLimit {
-    pub fn query(self) -> SearchQuery {
-        self.0
-    }
-}
