--- src/agdb/query_builder/select_values_ids.rs
@@ -1,9 +1,0 @@
-use crate::query::select_values_query::SelectValuesQuery;
-
-pub struct SelectValuesIds(pub SelectValuesQuery);
-
-impl SelectValuesIds {
-    pub fn query(self) -> SelectValuesQuery {
-        self.0
-    }
-}
