+++ src/agdb/query_builder/where_.rs
@@ -1,8 +1,8 @@
 use super::where_key::WhereKey;
 use super::where_logic_operator::WhereLogicOperator;
-use crate::query::condition::Condition;
-use crate::query::direction::Direction;
-use crate::query::edge_count_condition::EdgeCountCondition;
+use crate::query::query_condition::direction::Direction;
+use crate::query::query_condition::edge_count_condition::EdgeCountCondition;
+use crate::query::query_condition::QueryCondition;
 use crate::query::query_id::QueryId;
 use crate::query::query_ids::QueryIds;
 use crate::query::search_query::SearchQuery;
@@ -13,13 +13,13 @@ pub struct Where(pub SearchQuery);
 
 impl Where {
     pub fn distance(mut self, comparison: Comparison) -> WhereLogicOperator {
-        self.0.conditions.push(Condition::Distance(comparison));
+        self.0.conditions.push(QueryCondition::Distance(comparison));
 
         WhereLogicOperator(self.0)
     }
 
     pub fn edge(mut self) -> WhereLogicOperator {
-        self.0.conditions.push(Condition::Edge);
+        self.0.conditions.push(QueryCondition::Edge);
 
         WhereLogicOperator(self.0)
     }
@@ -27,7 +27,7 @@ impl Where {
     pub fn edge_count(mut self, comparison: Comparison) -> WhereLogicOperator {
         self.0
             .conditions
-            .push(Condition::EdgeCount(EdgeCountCondition {
+            .push(QueryCondition::EdgeCount(EdgeCountCondition {
                 comparison,
                 direction: Direction::Both,
             }));
@@ -38,7 +38,7 @@ impl Where {
     pub fn edge_count_from(mut self, comparison: Comparison) -> WhereLogicOperator {
         self.0
             .conditions
-            .push(Condition::EdgeCount(EdgeCountCondition {
+            .push(QueryCondition::EdgeCount(EdgeCountCondition {
                 comparison,
                 direction: Direction::From,
             }));
@@ -49,7 +49,7 @@ impl Where {
     pub fn edge_count_to(mut self, comparison: Comparison) -> WhereLogicOperator {
         self.0
             .conditions
-            .push(Condition::EdgeCount(EdgeCountCondition {
+            .push(QueryCondition::EdgeCount(EdgeCountCondition {
                 comparison,
                 direction: Direction::To,
             }));
@@ -64,7 +64,7 @@ impl Where {
     pub fn ids(mut self, ids: &[QueryId]) -> WhereLogicOperator {
         self.0
             .conditions
-            .push(Condition::Ids(QueryIds::Ids(ids.to_vec())));
+            .push(QueryCondition::Ids(QueryIds::Ids(ids.to_vec())));
 
         WhereLogicOperator(self.0)
     }
@@ -77,31 +77,31 @@ impl Where {
     }
 
     pub fn keys(mut self, names: &[DbKey]) -> WhereLogicOperator {
-        self.0.conditions.push(Condition::Keys(names.to_vec()));
+        self.0.conditions.push(QueryCondition::Keys(names.to_vec()));
 
         WhereLogicOperator(self.0)
     }
 
     pub fn node(mut self) -> WhereLogicOperator {
-        self.0.conditions.push(Condition::Node);
+        self.0.conditions.push(QueryCondition::Node);
 
         WhereLogicOperator(self.0)
     }
 
     pub fn not(mut self) -> Self {
-        self.0.conditions.push(Condition::Not);
+        self.0.conditions.push(QueryCondition::Not);
 
         self
     }
 
     pub fn not_beyond(mut self) -> Self {
-        self.0.conditions.push(Condition::NotBeyond);
+        self.0.conditions.push(QueryCondition::NotBeyond);
 
         self
     }
 
     pub fn where_(mut self) -> Self {
-        self.0.conditions.push(Condition::Where);
+        self.0.conditions.push(QueryCondition::Where);
 
         self
     }
