+++ src/agdb/query_builder/where_key.rs
@@ -1,7 +1,7 @@
 use super::where_::Where;
 use super::where_logic_operator::WhereLogicOperator;
-use crate::query::condition::Condition;
-use crate::query::key_value_condition::KeyValueCondition;
+use crate::query::query_condition::key_value_condition::KeyValueCondition;
+use crate::query::query_condition::QueryCondition;
 use crate::query::search_query::SearchQuery;
 use crate::Comparison;
 use crate::DbKey;
@@ -13,25 +13,27 @@ pub struct WhereKey {
 
 impl WhereKey {
     pub fn and(mut self) -> Where {
-        self.search.conditions.push(Condition::And);
+        self.search.conditions.push(QueryCondition::And);
 
         Where(self.search)
     }
 
     pub fn end_where(mut self) -> WhereLogicOperator {
-        self.search.conditions.push(Condition::EndWhere);
+        self.search.conditions.push(QueryCondition::EndWhere);
 
         WhereLogicOperator(self.search)
     }
 
     pub fn or(mut self) -> Where {
-        self.search.conditions.push(Condition::Or);
+        self.search.conditions.push(QueryCondition::Or);
 
         Where(self.search)
     }
 
     pub fn query(mut self) -> SearchQuery {
-        self.search.conditions.push(Condition::Keys(vec![self.key]));
+        self.search
+            .conditions
+            .push(QueryCondition::Keys(vec![self.key]));
 
         self.search
     }
@@ -39,7 +41,7 @@ impl WhereKey {
     pub fn value(mut self, comparison: Comparison) -> WhereLogicOperator {
         self.search
             .conditions
-            .push(Condition::KeyValue(KeyValueCondition {
+            .push(QueryCondition::KeyValue(KeyValueCondition {
                 key: self.key,
                 comparison,
             }));
